Refactor #36 - changed useCase package
