--- app/src/main/java/com/nlab/reminder/internal/feature/schedule/all/di/AllScheduleUseCaseModule.kt
@@ -1,50 +1,0 @@
-/*
- * Copyright (C) 2022 The N's lab Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.nlab.reminder.internal.feature.schedule.all.di
-
-import androidx.paging.PagingConfig
-import com.nlab.reminder.domain.common.schedule.ScheduleRepository
-import com.nlab.reminder.domain.feature.schedule.all.GetAllScheduleReportUseCase
-import com.nlab.reminder.domain.feature.schedule.all.UpdateScheduleCompleteUseCase
-import com.nlab.reminder.internal.feature.schedule.all.DefaultGetAllScheduleReportUseCase
-import com.nlab.reminder.internal.feature.schedule.all.DefaultUpdateScheduleCompleteUseCase
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ViewModelComponent
-import kotlinx.coroutines.Dispatchers
-
-/**
- * @author Doohyun
- */
-@Module
-@InstallIn(ViewModelComponent::class)
-class AllScheduleUseCaseModule {
-    @Provides
-    fun provideGetAllScheduleReportUseCase(
-        scheduleRepository: ScheduleRepository
-    ): GetAllScheduleReportUseCase = DefaultGetAllScheduleReportUseCase(
-        scheduleRepository,
-        pagingConfig = PagingConfig(pageSize = 10),
-        dispatcher = Dispatchers.Default
-    )
-
-    @Provides
-    fun provideUpdateScheduleCompleteUseCase(
-
-    ): UpdateScheduleCompleteUseCase = DefaultUpdateScheduleCompleteUseCase()
-}
